refactor: `MockClient` internal module
